+++ src/argument.h
@@ -15,7 +15,7 @@ public:
     static unsigned int num(const Nan::FunctionCallbackInfo<v8::Value> &info, int min = 0, int max = 0);
     static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, uint64* res, uint64 def);
     static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, std::vector<uint64> &res, uint64 def);
-    static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, std::vector<unsigned int> &res, uint64 def);
+    static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, std::vector<unsigned int> &res, unsigned int def);
     static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, unsigned int* res, unsigned int def);
     static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, int* res, int def);
     static unsigned int get(const Nan::FunctionCallbackInfo<v8::Value> &info, int arg, char** res, const char* def);
